feat(channel): introduce channel router
