+++ app/daemon/Cargo.toml
@@ -31,7 +31,7 @@ mcp = { workspace = true }
 protocol = { workspace = true }
 socket = { workspace = true }
 channel = { workspace = true }
-telegram = { workspace = true }
+channel_router = { workspace = true }
 compact_str = { workspace = true }
 futures-core = { workspace = true }
 futures-util = { workspace = true }
