+++ app/daemon/src/gateway/serve.rs
@@ -4,13 +4,11 @@
 //! through the shared dispatch path. A broadcast channel coordinates
 //! graceful shutdown across all subsystems.
 
-use crate::config::ChannelConfig;
 use crate::cron::{CronJob, CronScheduler};
-use crate::gateway::dispatch::AgentLock;
 use crate::gateway::{Gateway, GatewayHook};
 use crate::{DaemonConfig, loader};
 use anyhow::Result;
-use channel::{ChannelRouter, RoutingRule, parse_platform};
+use compact_str::CompactString;
 use model::ProviderManager;
 use runtime::Runtime;
 use std::path::{Path, PathBuf};
@@ -60,11 +58,9 @@ pub async fn serve_with_config(config: &DaemonConfig, config_dir: &Path) -> Resu
     let hf_endpoint = model::local::download::probe_endpoint().await;
     tracing::info!("using hf endpoint: {hf_endpoint}");
 
-    let locks = Arc::new(AgentLock::new());
     let runtime = Arc::new(runtime);
     let state = Gateway {
         runtime: Arc::clone(&runtime),
-        locks: Arc::clone(&locks),
         hf_endpoint: Arc::from(hf_endpoint),
     };
 
@@ -92,13 +88,22 @@ pub async fn serve_with_config(config: &DaemonConfig, config_dir: &Path) -> Resu
     ));
 
     // --- Channel transports ---
-    let router = build_router(&config.channels);
+    let router = channel_router::build_router(&config.channels);
     let router = Arc::new(router);
-    spawn_channels(&config.channels, &runtime, &locks, &router).await;
+    let rt = Arc::clone(&runtime);
+    let on_message = Arc::new(move |agent: CompactString, content: String| {
+        let rt = Arc::clone(&rt);
+        async move {
+            crate::gateway::dispatch::dispatch_send(&rt, &agent, &content)
+                .await
+                .map_err(|e| e.to_string())
+        }
+    });
+    channel_router::spawn_channels(&config.channels, router, on_message).await;
 
     // --- Cron scheduler ---
     let cron_dir = config_dir.join(crate::config::CRON_DIR);
-    spawn_cron(&cron_dir, &runtime, &locks, shutdown_tx.subscribe());
+    spawn_cron(&cron_dir, &runtime, shutdown_tx.subscribe());
 
     Ok(ServeHandle {
         socket_path: resolved_path,
@@ -117,85 +122,10 @@ fn bridge_shutdown(mut rx: broadcast::Receiver<()>) -> tokio::sync::oneshot::Rec
     orx
 }
 
-/// Build a [`ChannelRouter`] from the channel config entries.
-fn build_router(channels: &[ChannelConfig]) -> ChannelRouter {
-    let mut rules = Vec::new();
-    let mut default_agent = None;
-
-    for ch in channels {
-        let Ok(platform) = parse_platform(&ch.platform) else {
-            tracing::warn!("unknown platform '{}', skipping", ch.platform);
-            continue;
-        };
-        rules.push(RoutingRule {
-            platform,
-            channel_id: ch.channel_id.clone(),
-            agent: ch.agent.clone(),
-        });
-        if default_agent.is_none() {
-            default_agent = Some(ch.agent.clone());
-        }
-    }
-
-    ChannelRouter::new(rules, default_agent)
-}
-
-/// Connect and spawn channel loops for all configured channels.
-async fn spawn_channels(
-    channels: &[ChannelConfig],
-    runtime: &Arc<Runtime<ProviderManager, GatewayHook>>,
-    locks: &Arc<AgentLock>,
-    router: &Arc<ChannelRouter>,
-) {
-    for ch in channels {
-        let Ok(platform) = parse_platform(&ch.platform) else {
-            continue;
-        };
-
-        match platform {
-            channel::Platform::Telegram => {
-                let token = expand_env(&ch.bot_token);
-                let tg = telegram::TelegramChannel::new(token);
-                match channel::Channel::connect(tg).await {
-                    Ok(mut handle) => {
-                        let (tx, rx) = tokio::sync::mpsc::unbounded_channel();
-                        let sender = handle.sender();
-                        let rt = Arc::clone(runtime);
-                        let lk = Arc::clone(locks);
-                        let rr = Arc::clone(router);
-
-                        // Forward messages from ChannelHandle to the mpsc channel.
-                        tokio::spawn(async move {
-                            while let Some(msg) = handle.recv().await {
-                                if tx.send(msg).is_err() {
-                                    break;
-                                }
-                            }
-                        });
-
-                        tokio::spawn(crate::gateway::channel::channel_loop(
-                            rx, sender, rt, lk, rr,
-                        ));
-
-                        tracing::info!(platform = "telegram", "channel transport started");
-                    }
-                    Err(e) => {
-                        tracing::error!(platform = "telegram", "failed to connect channel: {e}");
-                    }
-                }
-            }
-            _ => {
-                tracing::warn!(platform = %ch.platform, "unsupported channel platform");
-            }
-        }
-    }
-}
-
 /// Load cron entries and start the scheduler.
 fn spawn_cron(
     cron_dir: &Path,
     runtime: &Arc<Runtime<ProviderManager, GatewayHook>>,
-    locks: &Arc<AgentLock>,
     shutdown: broadcast::Receiver<()>,
 ) {
     let entries = match loader::load_cron_dir(cron_dir) {
@@ -221,16 +151,12 @@ fn spawn_cron(
 
     let scheduler = CronScheduler::new(jobs);
     let rt = Arc::clone(runtime);
-    let lk = Arc::clone(locks);
 
     scheduler.start(
         move |job| {
             let rt = Arc::clone(&rt);
-            let lk = Arc::clone(&lk);
             async move {
-                match crate::gateway::dispatch::dispatch_send(&rt, &lk, &job.agent, &job.message)
-                    .await
-                {
+                match crate::gateway::dispatch::dispatch_send(&rt, &job.agent, &job.message).await {
                     Ok(response) => {
                         tracing::info!(
                             job = %job.name,
@@ -247,14 +173,4 @@ fn spawn_cron(
         },
         shutdown,
     );
-}
-
-/// Expand `${ENV_VAR}` patterns in a string. Returns the original if not found.
-fn expand_env(s: &str) -> String {
-    if s.starts_with("${") && s.ends_with('}') {
-        let var = &s[2..s.len() - 1];
-        std::env::var(var).unwrap_or_else(|_| s.to_owned())
-    } else {
-        s.to_owned()
-    }
 }
